solution Error  servitor confirmer configuration en “Capacitate Compartment

+++ APP.Eds/APP.Eds/Models/CompartimentCapacity/CompartimentCapacityModel.cs
@@ -4,5 +4,5 @@ public class CompartimentCapacityModel
 {
     public int IdCompartiment { get; set; }
     public int IdCapacity { get; set; }
-    public byte? Default { get; set; }
+    public double? Default { get; set; }
 }
